Rename compressor package to method
