+++ root/method/lzw.py
